requirements.txt
asgiref==3.8.1
Django==5.2
django-autoslug==1.9.9
django-debug-toolbar==5.1.0
django-widget-tweaks==1.5.0
pillow==11.2.1
psycopg==3.2.6
python-dotenv==1.1.0
sqlparse==0.5.3
